Add scripts to create PLINK files
The create_auxiliary_files.Rmd script creates a text file with subject IDs and a text file with sex information.
The create_plink_files.sh script creates PLINK files filtering for our selected subject IDs, for variants with an imputation score (R2) > 0.3, for variants with a MAF > 0.001, and with sex information. The script also creates LD files and a subsequent version of the PLINK files without rsIDs.